refactor: 자산 연동 관련 컴포넌트 파일 구조 정리
