+++ src/views/signup/components/asset/AssetSelectComponent.vue
