+++ src/views/signup/components/asset/ConnectModal.vue
