+++ src/views/signup/components/character/CharacterCard.vue
