+++ src/views/signup/components/character/CharacterDetailModal.vue
